+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Active Directory's `samAccountName` generation can now be customized ([#454]).
+
+[#454]: https://github.com/stackabletech/secret-operator/pull/454
+
 ## [24.7.0] - 2024-07-24
 
 ### Added